Connected k-rate AudioParams must get input values (1 of N)
When an AudioParam was selected to be k-rate, any inputs to the
AudioParam were ignored, so the AudioParam only contained the effect
of automation events.  This is incorrect.  Any inputs should be
reflected in the AudioParam values.

The case of connected AudioParams with a-rate automation was working fine.

Handles ConstantSource, Gain, and StereoPanner since these have very
similar AudioParam processing with only one simple AudioParam.

Bug: 1015760
Change-Id: I29ec3d7b58248a96da911e03890b74f6fb1bf098