Add instrumentation facility to wptrunner
Add simple high-level instrumentation to record time spent in
different kinds of tasks, and particularly in different tests
and directories. This is based on explicit annotations in the
code to say which kind of task is currently running.

The output is intended to be compatible with flamegraph.pl to
generate visualisations of the time spent doing different tasks.